Moved enum export outside of route.ts
