+++ apps/balancer-tools/src/app/apr/(utils)/calculatePoolStats.ts
@@ -3,10 +3,19 @@ import * as balEmissions from "#/lib/balancer/emissions";
 import { Pool } from "#/lib/balancer/gauges";
 import { pools } from "#/lib/gql/server";
 
-import { PoolStatsData, PoolTokens, PoolTypeEnum } from "../api/route";
+import { PoolStatsData, PoolTokens } from "../api/route";
 import { getBALPriceByRound } from "./getBALPriceByRound";
 import { getPoolRelativeWeight } from "./getRelativeWeight";
 import { Round } from "./rounds";
+
+export enum PoolTypeEnum {
+  PHANTOM_STABLE = "ComposableStable",
+  WEIGHTED = "Weighted",
+  GYROE = "GyroE",
+  STABLE = "Stable",
+  MetaStable = "MetaStable",
+  UNKNOWN = "FX",
+}
 
 const WEEKS_IN_YEAR = 52;
